Reimplement the color wheel as the color editor
A property is present in GtkColorChooserWidget called
"use-colorwheel", which is enabled by default, that will cause the
(not anymore deprecated) color wheel to be used as the single-color
editor (found when clicking the "Add" button below the color palette).
As a user, the exact same can be done in the color wheel compared to
the basic editor, but the color wheel has the hue, saturation and
brightness as an added bonus.